+++ infinite-tracing/src/test/java/com/newrelic/SpanEventSenderTest.java
@@ -13,12 +13,10 @@ import org.mockito.ArgumentMatchers;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
 
-import static com.newrelic.SpanEventSender.convert;
+import static com.newrelic.SpanConverterTest.buildSpanEvent;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -99,7 +97,7 @@ class SpanEventSenderTest {
 
         target.pollAndWrite();
 
-        verify(target).writeToObserver(observer, convert(spanEvent));
+        verify(target).writeToObserver(observer, SpanConverter.convert(spanEvent));
     }
 
     @Test
@@ -139,39 +137,6 @@ class SpanEventSenderTest {
 
         assertEquals(spanEvent, target.pollSafely());
     }
-
-    @Test
-    void convert_Valid() throws IOException {
-        SpanEvent spanEvent = buildSpanEvent();
-
-        V1.Span result = convert(spanEvent);
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        result.writeTo(baos);
-        V1.Span deserialized = V1.Span.parseFrom(baos.toByteArray());
-
-        assertEquals("abc123", deserialized.getTraceId());
-        assertEquals("abc123", deserialized.getIntrinsicsOrThrow("traceId").getStringValue());
-        assertEquals("my app", deserialized.getIntrinsicsOrThrow("appName").getStringValue());
-        assertEquals("value", deserialized.getIntrinsicsOrThrow("intrStr").getStringValue());
-        assertEquals(12345, deserialized.getIntrinsicsOrThrow("intrInt").getIntValue());
-        assertEquals(3.14, deserialized.getIntrinsicsOrThrow("intrFloat").getDoubleValue(), 0.00001);
-        assertTrue(deserialized.getIntrinsicsOrThrow("intrBool").getBoolValue());
-        assertFalse(deserialized.containsIntrinsics("intrOther"));
-    }
-
-    private SpanEvent buildSpanEvent() {
-        return SpanEvent.builder()
-                .appName("my app")
-                .putIntrinsic("traceId", "abc123")
-                .putIntrinsic("intrStr", "value")
-                .putIntrinsic("intrInt", 12345)
-                .putIntrinsic("intrFloat", 3.14)
-                .putIntrinsic("intrBool", true)
-                .putIntrinsic("intrOther", TestEnum.ONE)
-                .build();
-    }
-
-    private enum TestEnum { ONE }
 
     @Test
     void writeToObserver_RethrowsException() {
